[fix] rename serivice to service
